Migrate /team/* page group
improve

import type

Finalize

support isBookingPage

finalize

fix

Fix build error

fix type errors

finalize

fix embed pages
