+++ apps/web/app/future/apps/categories/[category]/page.tsx
@@ -2,6 +2,7 @@ import CategoryPage from "@pages/apps/categories/[category]";
 import { Prisma } from "@prisma/client";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
+import { type GetServerSidePropsContext } from "next";
 import { notFound } from "next/navigation";
 import z from "zod";
 
@@ -38,8 +39,8 @@ const querySchema = z.object({
   category: z.nativeEnum(AppCategories),
 });
 
-const getPageProps = async ({ params }: { params: Record<string, string | string[]> }) => {
-  const p = querySchema.safeParse(params);
+const getPageProps = async (context: GetServerSidePropsContext) => {
+  const p = querySchema.safeParse(context.params);
 
   if (!p.success) {
     return notFound();
@@ -66,6 +67,5 @@ const getPageProps = async ({ params }: { params: Record<string, string | string
   };
 };
 
-// @ts-expect-error getData arg
-export default WithLayout({ getData: getPageProps, Page: CategoryPage })<"P">;
+export default WithLayout({ getData: getPageProps, Page: CategoryPage, getLayout: null })<"P">;
 export const dynamic = "force-static";
