+++ apps/web/app/future/team/[slug]/embed/page.tsx
@@ -1,0 +1,9 @@
+import LegacyPage from "@pages/team/[slug]/embed";
+import withEmbedSsrAppDir from "app/WithEmbedSSR";
+import { WithLayout } from "app/layoutHOC";
+
+import { getData } from "../page";
+
+const getEmbedData = withEmbedSsrAppDir(getData);
+
+export default WithLayout({ getLayout: null, getData: getEmbedData, Page: LegacyPage })<"P">;
