+++ apps/web/pages/team/[slug]/embed.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import withEmbedSsr from "@lib/withEmbedSsr";
 
 import { getServerSideProps as _getServerSideProps } from "../../team/[slug]";
